requirements.txt
# Flask and Extensions
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.1
Flask-Session==0.6.0
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
email-validator


# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Google Drive
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Production Server
gunicorn==21.2.0
gevent==24.2.1

# Security
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
redis==5.0.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0

# Required by Koyeb
Werkzeug>=2.3.7
click>=8.1.7
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.3

# Performance and Monitoring
prometheus-client==0.19.0
sentry-sdk[flask]==1.40.3